requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.1
fastapi==0.117.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
rich==14.1.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.43
SQLAlchemy-Utils==0.42.0
starlette==0.48.0
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
psycopg2-binary
bcrypt==4.1.2
passlib[bcrypt]
python-jose[cryptography]
email-validator
